1755 refactoring
