+++ .gitignore
@@ -6,13 +6,13 @@
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
-# User-specific stuff
+# ru.rrenat358.persist.User-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
 .idea/**/usage.statistics.xml
 .idea/**/dictionaries
 .idea/**/shelf
-# AWS User-specific
+# AWS ru.rrenat358.persist.User-specific
 .idea/**/aws.xml
 # Generated files
 .idea/**/contentModel.xml
