+++ Lesson02-Spring-Basics/VideoPoint.Lesson02.md
@@ -2,7 +2,7 @@
 VideoPoint
 ---
 
-00-43-46 == UserService  
+00-43-46 == ru.rrenat358.UserService  
 00-53-00 == pom.xml   
 01-00-00 == spring-context.xml   
 01-07-00 == Run mvpApp in console
