+++ Lesson02-Spring-Basics/src/main/java/ru/rrenat358/UserService.java
@@ -1,3 +1,5 @@
+package ru.rrenat358;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
