+++ Lesson02-Spring-Basics/src/main/java/ru/rrenat358/circular/SecondBean.java
@@ -1,4 +1,4 @@
-package circular;
+package ru.rrenat358.circular;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
