+++ Lesson02-Spring-Basics/src/main/java/ru/rrenat358/persist/UserRepository.java
@@ -1,10 +1,6 @@
 package ru.rrenat358.persist;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
 
 public interface UserRepository {
